test: yaml-parser: Test dictionary items ordering
While YAML specifies that mappings are unordered, the Raspberry Pi IPA
relies on elements being ordered as in the YAML data. To replace the
dependency on boost with the YamlParser class, we thus need to guarantee
that the order is preserved. Update the corresponding unit test to
ensure this. The test currently fails at the YamlParser doesn't
correctly preserve the order, this will be fixed by the next commit.

This commit should be reverted when the Raspberry Pi IPA updates to a
new tuning data format and drops support for the old format.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Naushir Patuck <naush@raspberrypi.com>
Tested-by: Naushir Patuck <naush@raspberrypi.com>