Documentation: Align Doxyfile.in style with doxygen
Align the style of Doxyfile.in with the `doxygen -g` output, to minimize
diffs in the future when updating the file with `doxygen -u`. This
includes

- Replacing tabs with spaces
- Replacing += with \ to continue lines
- Dropping the \ on the last line of an option

Also update the Doxygen version number to indicate that the file has
been updated to v1.9.5, even if no modification to the options in use
were needed.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Umang Jain <umang.jain@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>