Documentation: Strip default options and comments from Doxyfile.in
Doxyfile.in was generated using `doxygen -g` with Doxygen v1.8.14. This
has produced a file with all options supported by Doxygen at the time,
and a handful of them have then been modified to support libcamera's
needs, with most options retaining their default.

Some of the default options have been deprecated in newer Doxygen
releases, which led to warnings when generating the documentation. The
usual process to fix this is to update the configuration file with
`doxygen -u`. This creates a big diff due to the large number of changes
to default options, which isn't quite nice. Instead, let's just drop all
the default options and the comments, to produce a small and more
readable Doxyfile.in.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Umang Jain <umang.jain@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>